+++ src/Components/TicketFormConfigDialog.tsx
@@ -1,10 +1,9 @@
 import React from "react";
 import * as Scrivito from "scrivito";
 
-import i18n from "../config/i18n";
 import { Keyable } from "../utils/types";
 import { TenantContextProvider, useTenantContext } from "./TenantContextProvider";
-import { SortableContainer } from "./SortableContainer";
+import { SortableObjList } from "./TicketFormConfigDialog/SortableObjList";
 
 function transformPropertiesToArray(input: Keyable, uiSchema: Keyable) {
   const fieldList = Object.entries(input).map(([key, schema]: [string, Keyable]) => ({
@@ -39,18 +38,13 @@ function fromPropertiesDefinitionToSchema(inputList: Keyable[]) {
   };
 }
 
-enum TicketFormConfigDialogMode {
-  CREATE = "create",
-  DETAILS = "details"
-}
-
-const TicketFormConfigDialog = ({ object, mode }: { object: Scrivito.Obj, mode: TicketFormConfigDialogMode }) => (
+const TicketFormConfigDialog = ({ object }: { object: Scrivito.Obj }) => (
   <TenantContextProvider>
-    <TicketFormConfigDialogContent object={object} mode={mode} />
+    <TicketFormConfigDialogContent object={object} />
   </TenantContextProvider>
 );
 
-const TicketFormConfigDialogContent = ({ object, mode }: { object: Scrivito.Obj, mode: TicketFormConfigDialogMode }) => {
+const TicketFormConfigDialogContent = ({ object }: { object: Scrivito.Obj }) => {
   const [orderedObjs, setOrderedObjs] = React.useState<Keyable[]>([]);
   const { prepareTicketSchema, instance } = useTenantContext();
   const [ticketSchema, setTicketSchema] = React.useState<Keyable | null>();
@@ -94,74 +88,14 @@ const TicketFormConfigDialogContent = ({ object, mode }: { object: Scrivito.Obj,
   };
 
   return (
-    <div className="scrivito_modal_large scrivito_show p-0 shadow-none">
-      <div className="scrivito_modal_body">
-        <SortableObjList
-          objs={orderedObjs}
-          onSortEnd={onSortEnd}
-          updateField={updateField}
-          mode={mode}
-        />
-      </div>
-    </div>
-  );
-};
-
-function SortableObjList({
-  objs,
-  onSortEnd,
-  updateField,
-  mode,
-}) {
-  return (
-    <div id="scrivito_obj_sorting_sortable">
-      <SortableContainer
-        ids={objs.map((obj: Keyable) => obj.name)}
-        items={objs}
+    <div className="scrivito_detail_content">
+      <SortableObjList
+        objs={orderedObjs}
         onSortEnd={onSortEnd}
-        disabled={false}
-        useDragHandle={null}
-      >
-        {objs.map((obj) => (
-          <SortableObjListItem
-            key={obj.name}
-            obj={obj}
-            updateField={updateField}
-            mode={mode}
-          />
-        ))}
-      </SortableContainer>
+        updateField={updateField}
+      />
     </div>
-  );
-}
-
-const SortableObjListItem = ({ obj, updateField, mode }) => {
-  return (
-    <li>
-      <div>
-        {i18n.t("TicketFormConfigDialog.fieldname")}
-        {" "}
-        {obj.title || obj.name}
-      </div>
-      {!obj["ui:regular"] && (
-        <div>
-          <label>
-            <input
-              type="checkbox"
-              name="show"
-              checked={obj.show}
-              onChange={(_event) => {
-                updateField(obj, { show: !obj.show });
-              }}
-            />
-            {" "}
-            {mode === TicketFormConfigDialogMode.CREATE && i18n.t("TicketFormConfigDialog.showInCreateForm")}
-            {mode === TicketFormConfigDialogMode.DETAILS && i18n.t("TicketFormConfigDialog.showInDetails")}
-          </label>
-        </div>
-      )}
-    </li>
   );
 };
 
-export { TicketFormConfigDialog, TicketFormConfigDialogMode };
+export { TicketFormConfigDialog };
