+++ src/Components/TicketFormConfigDialog/SortableContainer.tsx
@@ -16,7 +16,7 @@ import {
 } from "@dnd-kit/sortable";
 import { CSS } from "@dnd-kit/utilities";
 
-import { Keyable } from "../utils/types";
+import { Keyable } from "../../utils/types";
 
 export function SortableContainer({
   items,
