+++ src/Objs/TicketFormConfiguration/index.ts
@@ -1,0 +1,2 @@
+import "./TicketFormConfigurationEditingConfig";
+import "./TicketFormConfigurationObjClass";
