+++ src/Objs/TicketPage/index.ts
@@ -1,0 +1,3 @@
+import "./TicketPageComponent";
+import "./TicketPageEditingConfig";
+import "./TicketPageObjClass";
