+++ src/Objs/index.ts
@@ -1,5 +1,2 @@
-import "./TicketPage/TicketPageComponent";
-import "./TicketPage/TicketPageEditingConfig";
-import "./TicketPage/TicketPageObjClass";
-import "./TicketFormConfiguration/TicketFormConfigurationEditingConfig";
-import "./TicketFormConfiguration/TicketFormConfigurationObjClass";
+import "./TicketPage";
+import "./TicketFormConfiguration";
