+++ src/config/locales/de.json
@@ -74,8 +74,7 @@
   "Status": "Status",
   "TicketFormConfigDialog": {
     "fieldname": "Feldname:",
-    "showInCreateForm": "Im Formular anzeigen",
-    "showInDetails": "In Details anzeigen"
+    "show": "anzeigen"
   },
   "Ticket": {
     "labels": {
